moved useSignIn hook to /hooks, added it to all pages
