+++ pages/application.tsx
@@ -1,15 +1,14 @@
 import React, { ReactElement } from 'react';
-import { useSession } from 'next-auth/client';
-import Router, { useRouter } from 'next/router';
 import { QuestionType } from '../common/types';
 import { EXAMPLE_QUESTIONS } from '../common/constants';
 import ShortTextQuestion from '../components/ShortTextQuestion';
 import LongTextQuestion from '../components/LongTextQuestion';
 import CheckboxesQuestion from '../components/CheckboxesQuestion';
 import DropdownQuestion from '../components/DropdownQuestion';
+import { useSignIn } from '../hooks/useSignIn';
 
 const Application = (): ReactElement => {
-  const session = useSignIn();
+  useSignIn();
   return (
     <>
       <h1>Application Page</h1>
@@ -31,14 +30,3 @@ const Application = (): ReactElement => {
   );
 };
 export default Application;
-
-export function useSignIn() {
-  const [session, loading] = useSession();
-  const { pathname } = useRouter();
-
-  if (session) {
-    return session;
-  } else if (!loading && pathname !== '/login') {
-    Router.push('/login');
-  }
-}
