+++ pages/dashboard.tsx
@@ -2,8 +2,10 @@ import React, { ReactElement } from 'react';
 import { getApplicationStatus, getWelcome } from '../common/apiClient';
 import { APPLY_BY_DATE } from '../common/constants';
 import useSWR from 'swr';
+import { useSignIn } from '../hooks/useSignIn';
 
 const Dashboard = (): ReactElement => {
+  useSignIn();
   const { data: welcome } = useSWR('/api/v1/welcome', getWelcome);
   const { data: status } = useSWR('/api/v1/status', getApplicationStatus);
