+++ pages/home.tsx
@@ -1,6 +1,8 @@
 import React, { ReactElement } from 'react';
+import { useSignIn } from '../hooks/useSignIn';
 
 const Home = (): ReactElement => {
+  useSignIn();
   return <div>Home Page</div>;
 };
